remove DeviceSnapshot controller/view
